requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.3
statsmodels==0.14.0

# Machine Learning
tensorflow==2.13.0
torch==2.0.1
transformers==4.30.2
pytorch-lightning==2.0.6
xgboost==1.7.6
optuna==3.3.0

# Web and API
flask==2.2.5
flask-socketio==5.3.4
flask-restful==0.3.10
flask-sqlalchemy==3.0.5
dash==2.13.0
plotly==5.16.1

# Data sources and APIs
python-binance==1.0.17
ccxt==3.1.54
requests==2.31.0
websocket-client==1.6.1
pycoingecko==3.1.0

# Technical analysis
ta==0.10.2
ta-lib==0.4.26
pandas-ta==0.3.14b0

# Financial utilities
empyrical==0.5.5
pyfolio==0.9.2
alphalens==0.4.0

# Natural Language Processing
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1
vaderSentiment==3.3.2

# Database
sqlalchemy==2.0.20
pymongo==4.5.0
redis==4.6.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2
loguru==0.7.0
pydantic==1.10.8
typing-extensions==4.5.0
fastapi==0.95.2
uvicorn==0.22.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
bokeh==3.2.2

# Ollama LLM Integration
ollama==0.1.6

# Installation
# To install the packages, run the following command:
python install_essential.py
